amhello: substitute variables values using make's dollar sign syntax
In Makefiles, one can substitute a variable's value using a dollar sign
syntax which can use parentheses or braces: either $(foo) or ${foo}.

The other way, @foo@ worked because the M4 PKG_CHECK_MODULES function,
uses AC_ARG_VAR to create the _CFLAGS and _LIBS variables, and that
uses AC_SUBST which replaces instances of @foo@ in input files with
the value that the shell variable variable.

Note that CMake uses the form of ${foo} to reference variables.

Refs:
https://www.gnu.org/software/make/manual/html_node/Reference.html
https://github.com/pkgconf/pkgconf/blob/master/pkg.m4
https://www.gnu.org/software/autoconf/manual/autoconf-2.64/html_node/Setting-Output-Variables.html
https://cmake.org/cmake/help/latest/manual/cmake-language.7.html#id23
https://earthly.dev/blog/makefile-variables/
https://www.gnu.org/savannah-checkouts/gnu/autoconf/manual/autoconf-2.72/html_node/Makefile-Substitutions.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>