amhello: prepare SOURCES, CFLAGS and LDADD variables to be expanded
Makefiles has a “line-based” syntax in which the newline character
is special and marks the end of a statement.

But for readability we can format our makefiles by adding newlines
into the middle of a statement: we do this by escaping the internal
newlines with a backslash (\) character.

This way when adding new files, flags or dependencies is much easier.

Refs:
https://drmikehenry.com/gnu-make-line-continuations/
https://www.gnu.org/software/make/manual/html_node/Splitting-Lines.html
https://www.gnu.org/software/automake/manual/html_node/Program-and-Library-Variables.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>