tribe: use the same name for action and UI element
It is confusing, but when converting to xml the reason will become
self-evident. It is a matter of practicality. This is because there is a
trick GtkUIManager allow us to do. According to the documentation [1],
"if a name is not specified, it defaults to the action", so, if we only set
the value for action, it will use that for name and voila. Works like magic.
It may be a little confusing at first but the end result is cleaner.

Refs:
[1] - https://docs.gtk.org/gtk3/class.UIManager.html
https://docs.gtk.org/gtk3/method.UIManager.add_ui.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>