tribe: migrate menu code to xml
This defines the menu in xml and let GtkUIManager
build the menu for us from the xml string.

One reason for doing this was to approximate the gtk
code from the code used on the Windows side, which
defines a menu in a resource (.rc) file.

Another reason was to reduce the usage and thus the dependency
on GtkUIManager, as it was deprecated in GTK+ 3.10. We want
to leave the door open to migrate to version 3 of GTK in the
future.

A third reason is that this will make possible to move from
string into a totally separated and dedicated file, like
on Windows with resource (.rc) files.

I have been testing this for a couple of days, and seems to
be working fine.

Refs:
https://docs.gtk.org/gtk3/class.UIManager.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>