tribe: move sqlite3 files up one level
This moves the sqlite3 source files out of amalgamation directory,
and put them directly in the sqlite directory one level up.

This is because the fact sqlite3.c is a concatenation of over 100 separate
source files is not relevant and having "amalgamation" in the include
path is not very common as amalgamation is not a standard/common
directory name and is hard to type.

So, we are done with the amalgamation directory.

Refs:
https://www.sqlite.org/amalgamation.html

Signed-off-by: Thiago Farina <tfransosi@gmail.com>