c: add a small program to read a png image
This code was mostly copied from https://github.com/ShadowsocksR-Live/zbar/blob/master/examples/scan_image.c.

I want to read bar code and qr code images with it. That is the main goal.

It does not do much at the moment. For now, it just knows how to
get the width and height of the image.

I got the version of the png library I have installed on my system with
the following command:

$ pkg-config --modversion libpng
1.2.54

$ pkg-config --cflags --libs libpng
-I/usr/include/libpng12 -lpng12

Signed-off-by: Thiago Farina <tfransosi@gmail.com>