c: add a small program to prompt user for his password
The main goal was to disable the output, so when the user
is typing the password nothing is displayed. After his is
done entering the password and presses Enter, we display
the password we captured and restore the keyboard output.

This is far from being perfect, there are many details
overlooked that can be improved.

Refs:
https://stackoverflow.com/questions/2634203/enter-password-in-c
https://stackoverflow.com/questions/1196418/getting-a-password-in-c-without-using-getpass-3
https://stackoverflow.com/questions/1754004/how-to-mask-password-in-c?rq=3
https://stackoverflow.com/questions/32396188/better-alternative-for-getpass-function?rq=3

Signed-off-by: Thiago Farina <tfransosi@gmail.com>