tribe: fix release of resources in list command
If we return after a failure in ab_init, we never get a change to call
ab_fini and close the database. If we return after ab_enum_contacts_v2 fails,
we don't get a change to free the contacts array if it was allocated and we
don't get a change to close the database. And we should be closing the
database to prevent corruption and freeing contacts array to avoid memory
leak (which can be huge if the array has thoundsands or millions of entries).
And this is where goto helps a lot and keeps the code really simple and clean.

Tested running the following:

>tribels
>tribecli list
>tribecli ls

Signed-off-by: Thiago Farina <tfransosi@gmail.com>