c: add valuable information about enums
I kept the same wording used by the .NET reference because they explained
very clearly the information about bit flags (fields) I was looking for.

I think enums are a very special and interesting part of the C language. Hence,
I making a copy of this valuable information and keeping it around.

I know the enum constants can be written as macros and group together by some
kind of prefix, but enums offer a way to make them in lists and provide a scope
for them.

Used together with hexadecimal values (possibly power of twos) they became
very powerful.

Refs:
https://learn.microsoft.com/en-us/cpp/c-language/c-enumeration-declarations?view=msvc-170
https://learn.microsoft.com/en-us/dotnet/csharp/language-reference/builtin-types/enum

Signed-off-by: Thiago Farina <tfransosi@gmail.com>