tribe: reduce lifetime, scope and visibility of dbname variable
Previously, the variable dbname was declared and defined outside
all functions with the storage-class specifier static. And what
that means? By being declared with static, it means it has storage
and a defined value for the duration of the program. The storage is
reserved and its stored value is initialized only once, before the
program starts up. Why is this important? It is important, because
that means we are keeping the filename of the sqlite database on
memory during the whole execution of the program. Which, in reality,
is not necessary, as only when opening the database we need to know
the name of the file.

So, what are we doing here? We are basically reducing its lifetime (which
is just the period during execution of the program in which a variable
or a function exists).

For the scope, the variable had file scope, as it appeared outside any
block or list of parameters and because of that it was accessible from
any place in the source file (translation unit) after its declaration.
And that is another reason for reducing its scope and thus its visibility,
as no other parts of the code and functions in that file need to access
or use this variable. That is why variables with file scope are often
called "global" or "external".

Now, by moving it into the ab_init function we are giving it automatic
storage duration (turning it into a "local variable") and what that means?
It means the variable will now have storage and a defined value only
within that block where it was defined. It also means a new storage will
be allocated each time the program reaches that location, and the
variable will lose its value (and storage) as soon as the program exits
that block. This is the behavior we want for this variable, as it is
needed only for (when) opening the database.

Variables declared at the external-level or with the static storage-class
specified have global lifetime, but since it was declared with the static
keyword it was visible only within the file (translation unit) in which it
was declared. The static keyword gave it internal-level, overriding the
default behavior.

As we can see from this lengthy and long explanation, a lot of thought and
consideration should be put when declaring a variable. Where and how it is
declared will determine its lifetime, scope, visibility and linkage.

Tested running the following:

>tribe
>tribels
>tribecli list
>tribecli ls

Refs:
https://learn.microsoft.com/en-us/cpp/c-language/lifetime?view=msvc-170
https://learn.microsoft.com/en-us/cpp/c-language/scope-and-visibility?view=msvc\
-170
https://learn.microsoft.com/en-us/cpp/c-language/linkage?view=msvc-170
https://learn.microsoft.com/en-us/cpp/c-language/c-storage-classes?view=msvc-17\
0
https://learn.microsoft.com/en-us/cpp/c-language/storage-class-specifiers-for-e\
xternal-level-declarations?view=msvc-170
https://learn.microsoft.com/en-us/cpp/c-language/storage-class-specifiers-for-i\
nternal-level-declarations?view=msvc-170

Signed-off-by: Thiago Farina <tfransosi@gmail.com>