tribe: fix the name of the variable rows_deleted
This renames it from rows_deleted to num_change. Why? Because the use of
past tense of the verb "to delete" made it sound as if it was a boolean
variable, asking "was it deleted?", but in fact it is a counter (integer)
variable, as sqlite3_changes function returns the NUMBER of rows modified,
inserted or deleted by the most recently completed INSERT, UDATE or DELETE
statement. So, with this variable what we are interested is in how many (NUMber
of) rows were CHANGEd, not just if it "changed", like it sounded before.

So, we have to be careful and thoughtful when naming variables as well.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>