tribe: call sqlite3_changes before entering the cleanup area
The label statement ("out") delimits the start of the cleanup area. We reserve
this are for releasing and freeing resources (memory) allocated earlier, and
thus, it is not the most suitable place for verifying the result of
a sqlite3_step call. So we are moving the call to sqlite3_changes function
out of this area, and placing it just right before the label statement begins,
as it seems to be the most fitted place for it.

To test it, I made a change in cmd_delete.c to force a call to
ab_delete_contact_v2, bypassing the verification done by ab_get_contact_by_id.
When tested with -1, 0 or other non-existant ids sqlite3_changes returned 0.
When tested with some existant id, sqlite3_changes returned 1.

Example (of how I tested under gdb):
>gdb --args ./path/to/tribecli del 0
>b ab_delete_contact_v2

Signed-off-by: Thiago Farina <tfransosi@gmail.com>