tribe: change the name of the errcode variable to scode
Why? Because these functions are using it to indicate success or failure,
not just a failure. When the function begins, this variable is
initialized to zero (0) and this value is used to represent success. So,
if no error occurs during the execution of the function, it is misleading
to have it named errcode, because its value will be one indicating
success. So, to make it clear it does not contain just an error code,
we are renaming it to scode ("status code").

Zero means success and a negative error means failure.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>