c: use the logical-negation operator to check for NULL pointer
The logical-negation operator produces the value 0 if its operand is
true (nonzero) and the value 1 if its operand is false (0) and the
operand can be a pointer value.

There are some ways to check a pointer for NULL. One is if (ptr == NULL),
so if (!ptr) makes the code more succinct and clear. It avoids the
opportunity for an accidental assignment if someone misses one "=" and
writes if (ptr = NULL). This has also the benefit of not having to depend
on the definition of the NULL macro, which is ((void *)0) in C and 0 in C++.

Also, checks like if (ptr != NULL), become if (ptr), which is much more
succinct, making the code line much shorter.

From now on, the forms if (!ptr) and if (ptr) are the preferred way for
checking for NULL pointers.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>