c: get rid of len variable in xstrdup()
The len variable became unnecessary when we switched from memcpy() to strcpy()
with commit 9a2d485 ("c: use 'strcpy' in 'f_strdup' function"). It was
necessary back then because memcpy() takes a size parameter while strcpy()
does not.

strlen() returns the number of characters in str, excluding the terminal null.

strcpy() copies strSrc, including the terminating null character, to the
location that's specified by strDst.

So, I read the calculation of malloc's size parameter as the number of
characters in the source string plus one for the terminating null character
multiplied by the size of the underlying data type used (char in this case)
to represent a string. And this is the formula used to allocate memory for
a a string, as in C a string is an array of characters (char). For other
data types, to allocate memory for an array, the formula is the same but
without the +1 (plus one), which is specific to arrays of chars because
of the terminating null character.

When debugging under gdb we didn't lose any ability to inspect
the size of the string.

The following works:

(gdb) p strlen(str)
(gdb) p strlen(str) + 1
(gdb) p (strlen(str) + 1) * sizeof(char)
(gdb) p sizeof(char) works

Signed-off-by: Thiago Farina <tfransosi@gmail.com>