tribe: avoid using redundant != or == in conditions
Currently, the code in ab.c and ab_contact.c does NULL pointer
verifications using "ptr != NULL" and "ptr == NULL" and even using Yoda
conditions.

But since the expression can be of a pointer type and the body of the if
statement (same for while and for statements) is executed if the value of
the expression is nonzero, this can be simplified to if (ptr) and if
(!ptr).

This has some benefits. It makes the code less verbose. Avoids an
accidental assignment if some misses one equal sign ("="), the use of Yoda
conditions was guarding against this though. And thus helps the code style
become consistent and conventional.

This is a port of commit f760e400c09d ("avoid using redundant != or == in conditions") done at the Windows code base [1].

Refs:
[1] - https://github.com/tfarina/tribe/commit/f760e400c09d217a9abc40a9be978464e769fa62

Signed-off-by: Thiago Farina <tfransosi@gmail.com>