tribe: check the result of xstrdup() in _db_enum_contacts()
Our implementation of the strdup() function (xstrdup) does not exit on
failure, but returns NULL if a failure happens. For that reason,
we should check the result on the callers instead of silently
ignoring it.

So, let's check the result of xstrdup() function and jump to the error
handling path to propagate the error code.

The bonus of this change is that on the Windows side, we already do this.
So, this reduces the differences between both code bases, putting them
closer.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>