tribe: fix build error when -Wdeclaration-after-statement is enabled
Enabling this warning together with -Werror causes the following:

../src/ab.c: In function ‘_db_enum_contacts’:
../src/ab.c:238:5: error: ISO C90 forbids mixed declarations and code [-Wdeclaration-after-statement]
     unsigned char const *psz = NULL;
          ^

Commit a4577b4ec4a1 ("tribe: fix pointer-sign errors in _db_enum_contacts
and ab_get_contact_by_id functions") introduced a "psz" variable in the
for loop. But the issue was introduced by 704d23d74adb ("tribe: store the
return value of sqlite3_step in rc variable"), because it reworked the
if statement to store the result of sqlite3_step in
a variable and that is when the problem really came in.

We fix this build error by moving the variable declaration to the
beginning of the for statement, before the assignment.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>