tribe: fix variable naming in ab_enum_contacts()
Currently, there are some naming issues with the parameters "pCount"
and "ppContacts" of ab_enum_contacts() function.

The main naming issue is that they look nothing like the rest of the
file, because they are using the Hungarian notation [1] and a different
naming style.

The Hungarian prefix [1] "p" is for "pointer to" and "pp" for
"pointer to pointer to". So, the first parameter "pCount" reads as
"pointer to Count", which is not great, I know. It is too generic.
The second parameter "ppContacts" reads as "pointer to pointer to Contacts"
and it is better named, but has the issues pointed out above. The second
problem is that both are using the lower camelCase naming style, which
differs from the predominant style used for variables in the file, the
snake_case.

So, in order to make them look like the rest of the file and keep the
style consistent, we removed the Hungarian prefixes, renamed them to
"num_contacts" and "contacts_dst" respectively and changed the way it
is written to the snake_case naming style.

Let's reserve the use of Hungarian notation and the lower camelCase naming
style for the code used for Windows.

Refs:
https://en.wikipedia.org/wiki/Hungarian_notation

Signed-off-by: Thiago Farina <tfransosi@gmail.com>