tribe: rename _db_get_row_count() function
The naming issue with this function is that, although it may look
like properly and well named (following all the rules), one might
be asking: the row count of which *table*?.

In the context of a database with a single table, that would be fine,
but that is rarely the case. And, although that is the case with our
database right now, we are looking at expanding it, adding more tables
in the future.

So, we need to specify from which table we want to know many rows it has.

We use the A/HC/LC pattern for naming this function and pretty much every
other function in this file, but it seems to be lacking the High Context (HC),
which is the primary subject or entity in focus.

Currently, it has a "_db", which is a module prefix, "get", which is the
verb for the Action (A) and indicates what the primary operation is and
the "row_count", which is the Low Context (LC) as it narrows down the scope.

By introducing a "contacts" after the verb, we are not only make it more
specific, but are providing more information about which table we are
retriving the row count from to the callers. Which is better and will
allow the introduction of more functions like this for other tables
in the future. The function now named as _db_get_contacts_row_count(),
reads as "sqlite database get the number of rows in the contacts table".

The A/HC/LC pattern is described and explained in much more details at [1].

Refs:
[1] - https://johnfarrier.com/a-practical-guide-for-naming-things/

Signed-off-by: Thiago Farina <tfransosi@gmail.com>