tribe: NULL pointer to sqlite unconditionally
Currently, we are only setting this pointer to NULL in case the call to
sqlite3_close() function succeeds. This seems right, but we want to do this
regardless of the result of this call.

Also, calling sqlite3_close() with a NULL pointer is harmless. I verified this
calling _db_close() twice. When passing NULL, it returned 0 (SQLITE_OK) as
result.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>