tribe: fix some error messages
When opening or closing the database, we are giving up to the user the name
of the technology used, but that is not necessary. So, let's just say
"database".

Also, to avoid being overly technical, let's switch from "Failed to" to
"Unable to". I think "Unable to" even reads better, as if the software
was saying "I was unable to perform this action...".

We are basically following the recommendations for Error Messages in
Windows 7, by Microsoft's UX Guidelines [1].

Refs:
[1] - https://learn.microsoft.com/en-us/windows/win32/uxguide/mess-error

Signed-off-by: Thiago Farina <tfransosi@gmail.com>