c: add another version of songinfo
This one is built on top of taglib library.

I installed this library with:

> sudo apt install libtagc0

Then, to figure out the pkg-config name:

> pkg-config --list-all | grep tag

To figure out the version:

> pkg-config --modversion taglib_c

From there I had all the information to feed CMake and get it working.

With this library, it is pretty easy to pull out all the information
I want from the mp3 file.

Refs:
https://github.com/taglib/taglib

Signed-off-by: Thiago Farina <tfransosi@gmail.com>