c: rename stringize.h to base.h
Why? stringize is a verb [0], and verbs represent
actions, commands "to do" something, "to make" something.
But in file names it does not make sense to use a verb,
because the action it is "saying" will not happen.
The #include directive will not "stringize" anything
when the preprocessor (cpp) includes the content of the
file in another file. That is why using a verb in file
names is misleading and should not be used. Also,
a header file (.h) or source file (.c) can contain
an enourmous amount of code, not just a single macro or
a function. So, using a verb is very restricting, it
would be like as if it was limitating the file to only
one function.

The verb stringize means "to convert to a string".

Refs:
[0] - https://www.wordsense.eu/stringize/

Signed-off-by: Thiago Farina <tfransosi@gmail.com>

c#	../../asm/hello/hello_intel_x86.o