hello: define a real bug-report address in a separate macro
This way, when changing the url value, the history change and
git blame will be clearer.

Also, make it point to a real bug-report address, where issues
can really be reported.

Same as we did at commit 564f7cd8c973("amhello: define a real
bug-report address in a separate macro").

Signed-off-by: Thiago Farina <tfransosi@gmail.com>