tribe: suggest another name for the output directory
We are suggesting the user creates a directory named "build",
to build the source code there, but I plan to have a directory
called "build" checked in. In autotools build, there is even
a directory commonly called "build-aux".

So, to avoid any potential conflicts in the future, let's suggest
another directory. "out" seems to be a very fitting name for this.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>