tribe: fix parameter name in ab_init() prototype
dbpath was renamed to db_dir by commit 36e0bc117cae ("tribe:
fix variable naming issues in ab_init()") in order to clarify
to the callers they should pass a path to a directory where
the database should be saved.

I forgot to update the function declaration in the header file,
so let's fix this to keep them in sync.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>