c: use subscript operator [] in ARRAY_SIZE macro
In commit 473407909239 ("learn/c: demonstrate E1[0] array syntax and
its pointer equivalents"), we showed the compiler turns E1[0] or [0]E1
into *(E1 + 0) or simply *E1.

But the macro operates in an array type and the syntax [0] helps to inform
the reader what calculation it is doing. As it is dividing the total size
of the array by the size of the first element to calculate how many items
are there.

This basically reverts commit 260bc31fa84e ("c: use pointer syntax in ARRAY_SIZE macro"), which provided no explanation or reasoning for the change.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>