tribe: stop using EXECUTABLE_OUTPUT_PATH variable
We introduced this variable and started using it
at commit bc5fcedb616e ("sab: place binaries for
all build types in the same directory"). Back then,
we just followed the code in doxygen without even
checking the documentation. Which was a mistake.
We should have checked it before using it.

CMake's documentation [0], says it should no longer
be used as of CMake 2.6, and instead we use use
RUNTIME_OUTPUT_DIRECTORY. So, there we go.

Now that we introduced the GNUInstallDirs module at
commit b1ca38025af2 ("tribe: make use of GNUInstallDirs module"),
we are in better position to fix this and set
RUNTIME_OUTPUT_DIRECTORY properly.

So, let's set it to ${CMAKE_BINARY_DIR}/${CMAKE_INSTALL_BINDIR}
and stop using the old EXECUTABLE_OUTPUT_PATH variable.

Refs:
[0] - https://cmake.org/cmake/help/cmake2.6docs.html#variable:EXECUTABLE_OUTPUT_PATH

Signed-off-by: Thiago Farina <tfransosi@gmail.com>