tribe: rename GTK_REQUIRED variable to gtk_min_version
With commit 37081ed13759 ("tribe: force pkg-config to check
for a minimum specific version of gtk+-2.0") we forced
pkg-config to require a minimum version of gtk+-2.0, that is
it, starting from x we require at least version y to be able
to compile. Because with this we can control better the
dependencies and have a more predictable build.

In commit d8e41eb25558 ("tribe: move the check for required
packages to the top-level CMakeLists.txt file"), we moved
this pkg-config check into the top-level CMakeLists.txt file.

So, let's rename GTK_REQUIRED variable to gtk_min_version, to
help clarify that it is the minimum required version of gtk
library and avoid any confusion with "GTK REQUIRED" parameters
of pkg_check_modules() command.

Signed-off-by: Thiago Farina <tfransosi@gmail.com>